REFACTOR tests class to 'tests' roots
