--- tests/Feature/ConfigTest.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Sfneal\PostOffice\Tests\Feature;
-
-use Sfneal\PostOffice\Tests\TestCase;
-
-class ConfigTest extends TestCase
-{
-    // todo: add test methods
-}
