+++ tests/MailableTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Sfneal\PostOffice\Tests\Feature\Mailables;
+namespace Sfneal\PostOffice\Tests;
 
 use Sfneal\PostOffice\Mailables\AbstractMailable;
 use Sfneal\PostOffice\Tests\Assets\InvoiceUnpaidMailable;
 use Sfneal\PostOffice\Tests\TestCase;
 use Sfneal\Users\Models\User;
 
-class MailableImplementationsTest extends TestCase
+class MailableTest extends TestCase
 {
     /**
      * @var User
