+++ tests/ViewsTest.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Sfneal\PostOffice\Tests;
+
+class ViewsTest extends TestCase
+{
+    // todo: add test methods
+}
